+++ qml/AccountView.qml
@@ -47,7 +47,6 @@ ColumnLayout {
         id: stack_layout
         Layout.fillWidth: true
         Layout.fillHeight: true
-        Layout.bottomMargin: constants.p1
         currentIndex: account_view.currentView
 
         OverviewView {